@@ -14,43 +14,43 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^8.2.5",
-    "@angular/cdk": "^8.2.0",
-    "@angular/common": "^8.2.5",
-    "@angular/compiler": "^8.2.5",
-    "@angular/core": "^8.2.5",
-    "@angular/forms": "^8.2.5",
-    "@angular/http": "^7.2.15",
-    "@angular/material": "^8.2.0",
-    "@angular/platform-browser": "^8.2.5",
-    "@angular/platform-browser-dynamic": "^8.2.5",
-    "@angular/router": "^8.2.5",
-    "core-js": "^3.2.1",
+    "@angular/animations": "^9.1.6",
+    "@angular/cdk": "^9.2.3",
+    "@angular/common": "^9.1.6",
+    "@angular/compiler": "^9.1.6",
+    "@angular/core": "^9.1.6",
+    "@angular/forms": "^9.1.6",
+    "@angular/http": "^7.2.16",
+    "@angular/material": "^9.2.3",
+    "@angular/platform-browser": "^9.1.6",
+    "@angular/platform-browser-dynamic": "^9.1.6",
+    "@angular/router": "^9.1.6",
+    "core-js": "^3.6.5",
     "hammerjs": "^2.0.8",
     "rxjs": "^6.5.3",
     "zone.js": "^0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.803.21",
-    "@angular/cli": "^8.1.1",
-    "@angular/compiler-cli": "^8.1.1",
-    "@types/hammerjs": "^2.0.36",
-    "@types/jasmine": "3.3.13",
-    "@types/node": "~12.6.8",
-    "codelyzer": "~5.1.0",
-    "jasmine-core": "~3.4.0",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "^4.2.0",
-    "karma-chrome-launcher": "~3.0.0",
+    "@angular-devkit/build-angular": "^0.901.5",
+    "@angular/cli": "^9.1.5",
+    "@angular/compiler-cli": "^9.1.6",
+    "@types/jasmine": "~3.5.10",
+    "@types/node": "~13.13.5",
+    "codelyzer": "~5.2.2",
+    "jasmine-core": "~3.5.0",
+    "jasmine-spec-reporter": "~5.0.2",
+    "karma": "^5.0.5",
+    "karma-chrome-launcher": "~3.1.0",
     "karma-cli": "~2.0.0",
-    "karma-coverage-istanbul-reporter": "^2.0.5",
-    "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.2",
+    "karma-coverage-istanbul-reporter": "^3.0.2",
+    "karma-jasmine": "~3.1.1",
+    "karma-jasmine-html-reporter": "^1.5.3",
     "karma-phantomjs-launcher": "^1.0.4",
     "phantomjs-prebuilt": "^2.1.16",
-    "protractor": "^5.4.2",
-    "ts-node": "~8.3.0",
-    "tslint": "~5.18.0",
-    "typescript": "~3.4.3"
+    "prettier": "^2.0.5",
+    "protractor": "^6.0.0",
+    "ts-node": "~8.10.1",
+    "tslint": "~6.1.2",
+    "typescript": "~3.8.3"
   }
 }
