@@ -1,39 +1,25 @@
-import {
-  Component,
-  Input,
-  OnInit,
-  ViewContainerRef,
-  ViewChild
-} from "@angular/core";
-import { FormControl } from "@angular/forms";
-import {
-  MatAutocompleteTrigger,
-  MatDialog,
-  MatDialogConfig,
-  MatDialogRef
-} from "@angular/material";
-import { Observable } from "rxjs";
-import { startWith } from "rxjs/operators";
-import { map } from "rxjs/operators";
+import { Component, Input, OnInit, ViewContainerRef, ViewChild } from '@angular/core';
+import { FormControl } from '@angular/forms';
+import { MatDialog, MatDialogConfig, MatDialogRef } from '@angular/material/dialog';
+import { MatAutocompleteTrigger } from '@angular/material/autocomplete';
+import { Observable } from 'rxjs';
+import { startWith } from 'rxjs/operators';
+import { map } from 'rxjs/operators';
 
-import { Project, Skill, Tag } from "../models";
-import { ResumeService } from "../resume.service";
-import { ConfirmDialogComponent } from "../confirm-dialog/confirm-dialog.component";
+import { Project, Skill, Tag } from '../models';
+import { ResumeService } from '../resume.service';
+import { ConfirmDialogComponent } from '../confirm-dialog/confirm-dialog.component';
 
 @Component({
-  selector: "app-project",
-  templateUrl: "./project.component.html"
+  selector: 'app-project',
+  templateUrl: './project.component.html'
 })
 export class ProjectComponent implements OnInit {
   @Input() projects: Array<Project>;
   @Input() skills: Array<Skill>;
   months: Array<string>;
 
-  constructor(
-    private dialog: MatDialog,
-    private viewContainerRef: ViewContainerRef,
-    public resumeService: ResumeService
-  ) {}
+  constructor(private dialog: MatDialog, private viewContainerRef: ViewContainerRef, public resumeService: ResumeService) {}
 
   ngOnInit() {
     this.projects = this.projects || [];
@@ -46,9 +32,7 @@ export class ProjectComponent implements OnInit {
     config.width = '75vw';
 
     const dialogRef = this.dialog.open(ProjectDialog, config);
-    dialogRef.componentInstance.skills = (this.skills || []).map(
-      s => (s ? s.name : '')
-    );
+    dialogRef.componentInstance.skills = (this.skills || []).map(s => (s ? s.name : ''));
 
     dialogRef.afterClosed().subscribe(result => {
       if (result) {
@@ -64,7 +48,7 @@ export class ProjectComponent implements OnInit {
   editProject(project: Project) {
     const config = new MatDialogConfig();
     config.viewContainerRef = this.viewContainerRef;
-    config.width = "75vw";
+    config.width = '75vw';
 
     const dialogRef = this.dialog.open(ProjectDialog, config);
     dialogRef.componentInstance.project = project;
@@ -74,9 +58,7 @@ export class ProjectComponent implements OnInit {
       if (result) {
         project.endDate = project.current ? null : project.endDate;
         project.tags = result.tags;
-        this.projects = result
-          ? this.resumeService.updateProjects(this.projects)
-          : this.resumeService.retrieveResume().projects;
+        this.projects = result ? this.resumeService.updateProjects(this.projects) : this.resumeService.retrieveResume().projects;
         this.sortProjects();
       }
     });
@@ -87,9 +69,7 @@ export class ProjectComponent implements OnInit {
     config.viewContainerRef = this.viewContainerRef;
 
     const dialogRef = this.dialog.open(ConfirmDialogComponent, config);
-    dialogRef.componentInstance.message = `Are you sure you want to remove your project ${
-      project.name
-    }?`;
+    dialogRef.componentInstance.message = `Are you sure you want to remove your project ${project.name}?`;
     dialogRef.afterClosed().subscribe(result => {
       if (result) {
         this.projects = this.resumeService.removeProject(project);
@@ -100,9 +80,7 @@ export class ProjectComponent implements OnInit {
   // Sort by start date
   sortProjects() {
     this.projects.sort((a, b) => {
-      return b.startDate.year - a.startDate.year
-        ? b.startDate.year - a.startDate.year
-        : b.startDate.month - a.startDate.month;
+      return b.startDate.year - a.startDate.year ? b.startDate.year - a.startDate.year : b.startDate.month - a.startDate.month;
     });
   }
 }
@@ -110,49 +88,41 @@ export class ProjectComponent implements OnInit {
 // Add new project dialog
 @Component({
   template: `
-    <h3 mat-dialog-title>{{editMode ? 'Edit ' : 'Add new '}}project</h3>
+    <h3 mat-dialog-title>{{ editMode ? 'Edit ' : 'Add new ' }}project</h3>
     <div mat-dialog-content fxLayout="column">
       <mat-form-field>
-        <input matInput
-          [(ngModel)]="project.name"
-          placeholder="Project name">
+        <input matInput [(ngModel)]="project.name" placeholder="Project name" />
       </mat-form-field>
       <mat-form-field>
-        <textarea matInput
-          rows="4"
-          [(ngModel)]="project.description"
-          placeholder="Project description">
-        </textarea>
+        <textarea matInput rows="4" [(ngModel)]="project.description" placeholder="Project description"> </textarea>
       </mat-form-field>
       <mat-form-field>
-        <input matInput
-          [(ngModel)]="project.imageUrl"
-          placeholder="Project logo url (Optional)">
+        <input matInput [(ngModel)]="project.imageUrl" placeholder="Project logo url (Optional)" />
       </mat-form-field>
       <mat-form-field>
-        <input matInput
-          [(ngModel)]="project.web"
-          placeholder="Project website (Optional)">
+        <input matInput [(ngModel)]="project.web" placeholder="Project website (Optional)" />
       </mat-form-field>
 
       <mat-form-field>
         <mat-chip-list #tagList>
           <mat-chip *ngFor="let tag of tags" removable="true" (remove)="removeTag(tag)">
-            <mat-checkbox [checked]="tag.highlighted" (click)="toggleTagHighlight(tag)">{{tag.name}}</mat-checkbox>
+            <mat-checkbox [checked]="tag.highlighted" (click)="toggleTagHighlight(tag)">{{ tag.name }}</mat-checkbox>
             <mat-icon matChipRemove>cancel</mat-icon>
           </mat-chip>
-          <input matInput 
+          <input
+            matInput
             placeholder="Add tag..."
             [formControl]="tagControl"
             [matAutocomplete]="auto"
             [matChipInputFor]="tagList"
             [matChipInputAddOnBlur]="false"
             (optionSelected)="addTag($event)"
-            (matChipInputTokenEnd)="addTag($event)" />
+            (matChipInputTokenEnd)="addTag($event)"
+          />
         </mat-chip-list>
         <mat-autocomplete #auto="matAutocomplete" (optionSelected)="addTag($event)">
-          <mat-option *ngFor="let skill of filteredSkills | async" [value]="skill"  >
-              {{ skill }}
+          <mat-option *ngFor="let skill of filteredSkills | async" [value]="skill">
+            {{ skill }}
           </mat-option>
         </mat-autocomplete>
       </mat-form-field>
@@ -161,10 +131,14 @@ export class ProjectComponent implements OnInit {
         <label class="select-label">From</label>
         <div>
           <mat-select [(ngModel)]="project.startDate.year" placeholder="Year">
-            <mat-option *ngFor="let year of years" [value]="year"> {{year}} </mat-option>
+            <mat-option *ngFor="let year of years" [value]="year">
+              {{ year }}
+            </mat-option>
           </mat-select>
           <mat-select [(ngModel)]="project.startDate.month" placeholder="Month">
-            <mat-option *ngFor="let month of months" [value]="months.indexOf(month)"> {{month}} </mat-option>
+            <mat-option *ngFor="let month of months" [value]="months.indexOf(month)">
+              {{ month }}
+            </mat-option>
           </mat-select>
         </div>
       </div>
@@ -175,22 +149,31 @@ export class ProjectComponent implements OnInit {
         <label class="select-label">To</label>
         <div fxLayout="row">
           <mat-select [(ngModel)]="project.endDate.year" placeholder="Year">
-            <mat-option *ngFor="let year of years" [value]="year"> {{year}} </mat-option>
+            <mat-option *ngFor="let year of years" [value]="year">
+              {{ year }}
+            </mat-option>
           </mat-select>
           <mat-select [(ngModel)]="project.endDate.month" placeholder="Month">
-            <mat-option *ngFor="let month of months" [value]="months.indexOf(month)"> {{month}} </mat-option>
+            <mat-option *ngFor="let month of months" [value]="months.indexOf(month)">
+              {{ month }}
+            </mat-option>
           </mat-select>
         </div>
       </div>
     </div>
     <div mat-dialog-actions>
-      <button mat-button color="primary" (click)="dialogRef.close()">Cancel</button>
-      <button mat-button color="primary" (click)="dialogRef.close({ project: project, tags: tags})">Save</button>
+      <button mat-button color="primary" (click)="dialogRef.close()">
+        Cancel
+      </button>
+      <button mat-button color="primary" (click)="dialogRef.close({ project: project, tags: tags })">
+        Save
+      </button>
     </div>
   `
 })
 export class ProjectDialog implements OnInit {
-  @ViewChild(MatAutocompleteTrigger, { static: false }) autoTrigger: MatAutocompleteTrigger;
+  @ViewChild(MatAutocompleteTrigger, { static: false })
+  autoTrigger: MatAutocompleteTrigger;
 
   public project: Project;
   public tags: Array<Tag>;
@@ -201,10 +184,7 @@ export class ProjectDialog implements OnInit {
   public filteredSkills: Observable<Array<string>>;
   public tagControl: FormControl = new FormControl();
 
-  constructor(
-    public dialogRef: MatDialogRef<ProjectDialog>,
-    private resumeService: ResumeService
-  ) {
+  constructor(public dialogRef: MatDialogRef<ProjectDialog>, private resumeService: ResumeService) {
     this.project = {
       name: '',
       description: '',
@@ -223,9 +203,7 @@ export class ProjectDialog implements OnInit {
   ngOnInit() {
     // Assume edit mode if name isn't blank
     this.editMode = this.project && this.project.name.length > 0;
-    this.tags = this.project.tags
-      ? this.project.tags.map(t => Object.assign({}, t))
-      : [];
+    this.tags = this.project.tags ? this.project.tags.map(t => Object.assign({}, t)) : [];
 
     // Skills for tag autocomplete
     this.filteredSkills = this.tagControl.valueChanges.pipe(
@@ -237,17 +215,12 @@ export class ProjectDialog implements OnInit {
   filter(val: string): string[] {
     return this.skills
       .filter(s => this.tags.filter(t => t.name === s).length === 0)
-      .filter(
-        s => (val ? s.toLowerCase().indexOf(val.toLowerCase()) === 0 : true)
-      );
+      .filter(s => (val ? s.toLowerCase().indexOf(val.toLowerCase()) === 0 : true));
   }
 
   addTag(event: any): void {
     const input = event.input;
-    const value = (event.option
-      ? event.option.value
-      : event.value || ''
-    ).trim();
+    const value = (event.option ? event.option.value : event.value || '').trim();
 
     if (value) {
       this.tags.push({ name: value, highlighted: false });
