@@ -1,10 +1,11 @@
 /* tslint:disable:no-unused-variable */
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-import { By } from '@angular/platform-browser';
-import { DebugElement } from '@angular/core';
 import { HttpModule } from '@angular/http';
 import { HttpClientModule } from '@angular/common/http';
-import { MatIconModule, MatProgressBarModule, MatDialogModule, MatCardModule } from '@angular/material';
+import { MatIconModule } from '@angular/material/icon';
+import { MatProgressBarModule } from '@angular/material/progress-bar';
+import { MatDialogModule } from '@angular/material/dialog';
+import { MatCardModule } from '@angular/material/card';
 
 import { ResumeService } from '../resume.service';
 import { SkillComponent } from '../skill/skill.component';
@@ -19,16 +20,10 @@ describe('SummaryComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [
-        SummaryComponent,
-        SkillComponent
-      ],
-      providers: [
-        { provide: ResumeService, useValue: {} }
-      ],
-      imports: [ MatIconModule, MatProgressBarModule, MatDialogModule, HttpModule, HttpClientModule, MatCardModule ]
-    })
-    .compileComponents();
+      declarations: [SummaryComponent, SkillComponent],
+      providers: [{ provide: ResumeService, useValue: {} }],
+      imports: [MatIconModule, MatProgressBarModule, MatDialogModule, HttpModule, HttpClientModule, MatCardModule]
+    }).compileComponents();
   }));
 
   beforeEach(() => {
