@@ -1,8 +1,6 @@
 /* tslint:disable:no-unused-variable */
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-import { By } from '@angular/platform-browser';
-import { DebugElement } from '@angular/core';
-import { MatDialogRef } from '@angular/material';
+import { MatDialogRef } from '@angular/material/dialog';
 
 import { ConfirmDialogComponent } from './confirm-dialog.component';
 
@@ -12,10 +10,9 @@ describe('ConfirmDialogComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ ConfirmDialogComponent ],
-      providers: [{ provide: MatDialogRef, useValue: {} },]
-    })
-    .compileComponents();
+      declarations: [ConfirmDialogComponent],
+      providers: [{ provide: MatDialogRef, useValue: {} }]
+    }).compileComponents();
   }));
 
   beforeEach(() => {
