@@ -1,10 +1,11 @@
 /* tslint:disable:no-unused-variable */
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-import { By } from '@angular/platform-browser';
-import { DebugElement } from '@angular/core';
-import { MatIconModule, MatDialogModule, MatListModule, MatCardModule } from '@angular/material';
+import { MatIconModule } from '@angular/material/icon';
+import { MatListModule } from '@angular/material/list';
+import { MatDialogModule } from '@angular/material/dialog';
+import { MatCardModule } from '@angular/material/card';
 
-import { ResumeService } from '../resume.service'
+import { ResumeService } from '../resume.service';
 import { EducationComponent } from './education.component';
 
 describe('EducationComponent', () => {
@@ -13,13 +14,10 @@ describe('EducationComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [
-        EducationComponent
-      ],
-      providers: [{ provide: ResumeService, useValue: {} },],
-      imports: [ MatIconModule, MatDialogModule, MatListModule, MatCardModule ]
-    })
-    .compileComponents();
+      declarations: [EducationComponent],
+      providers: [{ provide: ResumeService, useValue: {} }],
+      imports: [MatIconModule, MatDialogModule, MatListModule, MatCardModule]
+    }).compileComponents();
   }));
 
   beforeEach(() => {
