@@ -1,10 +1,12 @@
 /* tslint:disable:no-unused-variable */
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-import { By } from '@angular/platform-browser';
-import { DebugElement } from '@angular/core';
-import { MatIconModule, MatDialogModule, MatChipsModule, MatMenuModule, MatCardModule } from '@angular/material';
+import { MatIconModule } from '@angular/material/icon';
+import { MatMenuModule } from '@angular/material/menu';
+import { MatDialogModule } from '@angular/material/dialog';
+import { MatCardModule } from '@angular/material/card';
+import { MatChipsModule } from '@angular/material/chips';
 
-import { ResumeService } from '../resume.service'
+import { ResumeService } from '../resume.service';
 import { ProjectComponent } from './project.component';
 
 describe('ProjectComponent', () => {
@@ -13,11 +15,10 @@ describe('ProjectComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ ProjectComponent ],
-      providers: [{ provide: ResumeService, useValue: {} },],
-      imports: [ MatIconModule, MatDialogModule, MatChipsModule, MatMenuModule, MatCardModule ]
-    })
-    .compileComponents();
+      declarations: [ProjectComponent],
+      providers: [{ provide: ResumeService, useValue: {} }],
+      imports: [MatIconModule, MatDialogModule, MatChipsModule, MatMenuModule, MatCardModule]
+    }).compileComponents();
   }));
 
   beforeEach(() => {
